refactor(DeviceConnector/RPCClient): Move RPCClient file and add service
