+++ packages/device-connector/src/arduino-cli/client.ts
