+++ packages/device-connector/src/deviceConnector.ts
@@ -1,6 +1,7 @@
-import { RPCClient } from './cli-rpc/client'
+import { connectCli, registerDevices } from './arduino-cli/service'
+import { RPCClient } from './arduino-cli/client'
 import { MessageEvent, WebSocket } from 'ws'
-import { deployBinary, setDevices } from './devices/service'
+import { deployBinary } from './devices/service'
 import { openStream } from './deployment-server/connection'
 import logger from './util/logger'
 import { Signals } from 'close-with-grace'
@@ -8,22 +9,6 @@ import { Signals } from 'close-with-grace'
 export interface DeviceConnector {
   client: RPCClient
   socket: WebSocket
-}
-
-const connectCli = async (): Promise<RPCClient> => {
-  const client = await new RPCClient()
-  await client.init()
-  await client.createInstance()
-  await client.initInstance()
-
-  return client
-}
-
-const registerDevices = async (client: RPCClient): Promise<void> => {
-  await client.boardListWatch()
-  const devices = client.getDevices()
-
-  setDevices(devices)
 }
 
 export const createConnector = async (): Promise<DeviceConnector> => {
