+++ packages/device-connector/src/devices/service.ts
@@ -1,5 +1,5 @@
 import { Port } from 'arduino-cli_proto_ts/common/cc/arduino/cli/commands/v1/Port'
-import { RPCClient } from '../cli-rpc/client'
+import { RPCClient } from '../arduino-cli/client'
 import {
   fetchAllDeviceTypes,
   fetchDeviceType,
